renamed highlight-controller to result-word-highlight
