+++ frontend/src/ts/elements/result-word-highlight.ts
