+++ frontend/src/ts/test/test-logic.ts
@@ -14,7 +14,7 @@ import * as Focus from "./focus";
 import * as Funbox from "./funbox/funbox";
 import * as Keymap from "../elements/keymap";
 import * as ThemeController from "../controllers/theme-controller";
-import * as HighlightController from "../controllers/highlight-controller";
+import * as ResultWordHighlight from "../elements/result-word-highlight";
 import * as PaceCaret from "./pace-caret";
 import * as Caret from "./caret";
 import * as LiveWpm from "./live-wpm";
@@ -456,7 +456,7 @@ export function restart(options = {} as RestartOptions): void {
     }
   );
 
-  HighlightController.destroy();
+  ResultWordHighlight.destroy();
 }
 
 let rememberLazyMode: boolean;
